requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
bitsandbytes>=0.41.0
accelerate>=0.23.0
peft>=0.6.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-image>=0.20.0
pandas>=2.0.0
tqdm>=4.65.0
langchain>=0.0.300
autogen>=0.2.0
huggingface_hub>=0.17.0
pytesseract>=0.3.10
difflib>=0.1
PyPDF2>=3.0.0